backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.4

# Image processing
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.4.0

# AI / Face recognition
deepface==0.0.93
retina-face==0.0.17
tensorflow==2.16.1
tf-keras==2.16.0
scikit-learn

# HTTP / LLM
requests==2.32.3
groq==0.9.0
